+++ builtin/credential/aws/backend_test.go
@@ -1,12 +1,17 @@
-package awsec2
+package awsauth
 
 import (
 	"encoding/base64"
+	"encoding/json"
 	"fmt"
+	"io/ioutil"
+	"net/http"
 	"os"
 	"strings"
 	"testing"
 
+	"github.com/aws/aws-sdk-go/aws/session"
+	"github.com/aws/aws-sdk-go/service/sts"
 	"github.com/hashicorp/vault/helper/policyutil"
 	"github.com/hashicorp/vault/logical"
 	logicaltest "github.com/hashicorp/vault/logical/testing"
@@ -29,11 +34,12 @@ func TestBackend_CreateParseVerifyRoleTag(t *testing.T) {
 
 	// create a role entry
 	data := map[string]interface{}{
+		"auth_type":    "ec2",
 		"policies":     "p,q,r,s",
 		"bound_ami_id": "abcd-123",
 	}
 	resp, err := b.HandleRequest(&logical.Request{
-		Operation: logical.UpdateOperation,
+		Operation: logical.CreateOperation,
 		Path:      "role/abcd-123",
 		Storage:   storage,
 		Data:      data,
@@ -100,7 +106,7 @@ func TestBackend_CreateParseVerifyRoleTag(t *testing.T) {
 
 	// register a different role
 	resp, err = b.HandleRequest(&logical.Request{
-		Operation: logical.UpdateOperation,
+		Operation: logical.CreateOperation,
 		Path:      "role/ami-6789",
 		Storage:   storage,
 		Data:      data,
@@ -683,132 +689,6 @@ vSeDCOUMYQR7R9LINYwouHIziqQYMAkGByqGSM44BAMDLwAwLAIUWXBlk40xTwSw
 	}
 }
 
-func TestBackend_pathRole(t *testing.T) {
-	config := logical.TestBackendConfig()
-	storage := &logical.InmemStorage{}
-	config.StorageView = storage
-
-	b, err := Backend(config)
-	if err != nil {
-		t.Fatal(err)
-	}
-	_, err = b.Setup(config)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	data := map[string]interface{}{
-		"policies":     "p,q,r,s",
-		"max_ttl":      "2h",
-		"bound_ami_id": "ami-abcd123",
-	}
-	resp, err := b.HandleRequest(&logical.Request{
-		Operation: logical.CreateOperation,
-		Path:      "role/ami-abcd123",
-		Data:      data,
-		Storage:   storage,
-	})
-	if resp != nil && resp.IsError() {
-		t.Fatalf("failed to create role")
-	}
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	resp, err = b.HandleRequest(&logical.Request{
-		Operation: logical.ReadOperation,
-		Path:      "role/ami-abcd123",
-		Storage:   storage,
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if resp == nil || resp.IsError() {
-		t.Fatal("failed to read the role entry")
-	}
-	if !policyutil.EquivalentPolicies(strings.Split(data["policies"].(string), ","), resp.Data["policies"].([]string)) {
-		t.Fatalf("bad: policies: expected: %#v\ngot: %#v\n", data, resp.Data)
-	}
-
-	data["allow_instance_migration"] = true
-	data["disallow_reauthentication"] = true
-	resp, err = b.HandleRequest(&logical.Request{
-		Operation: logical.UpdateOperation,
-		Path:      "role/ami-abcd123",
-		Data:      data,
-		Storage:   storage,
-	})
-	if resp != nil && resp.IsError() {
-		t.Fatalf("failed to create role")
-	}
-	if err != nil {
-		t.Fatal(err)
-	}
-	resp, err = b.HandleRequest(&logical.Request{
-		Operation: logical.ReadOperation,
-		Path:      "role/ami-abcd123",
-		Storage:   storage,
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !resp.Data["allow_instance_migration"].(bool) || !resp.Data["disallow_reauthentication"].(bool) {
-		t.Fatal("bad: expected:true got:false\n")
-	}
-
-	// add another entry, to test listing of role entries
-	resp, err = b.HandleRequest(&logical.Request{
-		Operation: logical.UpdateOperation,
-		Path:      "role/ami-abcd456",
-		Data:      data,
-		Storage:   storage,
-	})
-	if resp != nil && resp.IsError() {
-		t.Fatalf("failed to create role")
-	}
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	resp, err = b.HandleRequest(&logical.Request{
-		Operation: logical.ListOperation,
-		Path:      "roles",
-		Storage:   storage,
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if resp == nil || resp.Data == nil || resp.IsError() {
-		t.Fatalf("failed to list the role entries")
-	}
-	keys := resp.Data["keys"].([]string)
-	if len(keys) != 2 {
-		t.Fatalf("bad: keys: %#v\n", keys)
-	}
-
-	_, err = b.HandleRequest(&logical.Request{
-		Operation: logical.DeleteOperation,
-		Path:      "role/ami-abcd123",
-		Storage:   storage,
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	resp, err = b.HandleRequest(&logical.Request{
-		Operation: logical.ReadOperation,
-		Path:      "role/ami-abcd123",
-		Storage:   storage,
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if resp != nil {
-		t.Fatalf("bad: response: expected:nil actual:%#v\n", resp)
-	}
-
-}
-
 func TestBackend_parseAndVerifyRoleTagValue(t *testing.T) {
 	// create a backend
 	config := logical.TestBackendConfig()
@@ -825,6 +705,7 @@ func TestBackend_parseAndVerifyRoleTagValue(t *testing.T) {
 
 	// create a role
 	data := map[string]interface{}{
+		"auth_type":    "ec2",
 		"policies":     "p,q,r,s",
 		"max_ttl":      "120s",
 		"role_tag":     "VaultRole",
@@ -901,6 +782,7 @@ func TestBackend_PathRoleTag(t *testing.T) {
 	}
 
 	data := map[string]interface{}{
+		"auth_type":    "ec2",
 		"policies":     "p,q,r,s",
 		"max_ttl":      "120s",
 		"role_tag":     "VaultRole",
@@ -966,6 +848,7 @@ func TestBackend_PathBlacklistRoleTag(t *testing.T) {
 
 	// create an role entry
 	data := map[string]interface{}{
+		"auth_type":    "ec2",
 		"policies":     "p,q,r,s",
 		"role_tag":     "VaultRole",
 		"bound_ami_id": "abcd-123",
@@ -1068,7 +951,7 @@ func TestBackend_PathBlacklistRoleTag(t *testing.T) {
 // needs to be set:
 // TEST_AWS_SECRET_KEY
 // TEST_AWS_ACCESS_KEY
-func TestBackendAcc_LoginAndWhitelistIdentity(t *testing.T) {
+func TestBackendAcc_LoginWithInstanceIdentityDocAndWhitelistIdentity(t *testing.T) {
 	// This test case should be run only when certain env vars are set and
 	// executed as an acceptance test.
 	if os.Getenv(logicaltest.TestEnvVar) == "" {
@@ -1156,6 +1039,7 @@ func TestBackendAcc_LoginAndWhitelistIdentity(t *testing.T) {
 
 	// Place the wrong AMI ID in the role data.
 	data := map[string]interface{}{
+		"auth_type":          "ec2",
 		"policies":           "root",
 		"max_ttl":            "120s",
 		"bound_ami_id":       "wrong_ami_id",
@@ -1164,7 +1048,7 @@ func TestBackendAcc_LoginAndWhitelistIdentity(t *testing.T) {
 	}
 
 	roleReq := &logical.Request{
-		Operation: logical.UpdateOperation,
+		Operation: logical.CreateOperation,
 		Path:      "role/" + roleName,
 		Storage:   storage,
 		Data:      data,
@@ -1183,6 +1067,7 @@ func TestBackendAcc_LoginAndWhitelistIdentity(t *testing.T) {
 	}
 
 	// Place the correct AMI ID, but make the AccountID wrong
+	roleReq.Operation = logical.UpdateOperation
 	data["bound_ami_id"] = amiID
 	data["bound_account_id"] = "wrong-account-id"
 	resp, err = b.HandleRequest(roleReq)
@@ -1210,7 +1095,7 @@ func TestBackendAcc_LoginAndWhitelistIdentity(t *testing.T) {
 		t.Fatalf("bad: expected error response: resp:%#v\nerr:%v", resp, err)
 	}
 
-	// Place the correct IAM Role ARN
+	// place the correct IAM role ARN
 	data["bound_iam_role_arn"] = iamARN
 	resp, err = b.HandleRequest(roleReq)
 	if err != nil || (resp != nil && resp.IsError()) {
@@ -1280,7 +1165,6 @@ func TestBackend_pathStsConfig(t *testing.T) {
 	config := logical.TestBackendConfig()
 	storage := &logical.InmemStorage{}
 	config.StorageView = storage
-
 	b, err := Backend(config)
 	if err != nil {
 		t.Fatal(err)
@@ -1289,7 +1173,6 @@ func TestBackend_pathStsConfig(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-
 	stsReq := &logical.Request{
 		Operation: logical.CreateOperation,
 		Storage:   storage,
@@ -1387,5 +1270,235 @@ func TestBackend_pathStsConfig(t *testing.T) {
 	}
 	if resp.Data["keys"] != nil {
 		t.Fatalf("no entries should be present")
+	}
+}
+
+func buildCallerIdentityLoginData(request *http.Request, roleName string) (map[string]interface{}, error) {
+	headersJson, err := json.Marshal(request.Header)
+	if err != nil {
+		return nil, err
+	}
+	requestBody, err := ioutil.ReadAll(request.Body)
+	if err != nil {
+		return nil, err
+	}
+	return map[string]interface{}{
+		"iam_http_request_method": request.Method,
+		"iam_request_url":         base64.StdEncoding.EncodeToString([]byte(request.URL.String())),
+		"iam_request_headers":     base64.StdEncoding.EncodeToString(headersJson),
+		"iam_request_body":        base64.StdEncoding.EncodeToString(requestBody),
+		"request_role":            roleName,
+	}, nil
+}
+
+// This is an acceptance test.
+// If the test is NOT being run on an AWS EC2 instance in an instance profile,
+// it requires the following environment variables to be set:
+// TEST_AWS_ACCESS_KEY_ID
+// TEST_AWS_SECRET_ACCESS_KEY
+// TEST_AWS_SECURITY_TOKEN or TEST_AWS_SESSION_TOKEN (optional, if you are using short-lived creds)
+// These are intentionally NOT the "standard" variables to prevent accidentally
+// using prod creds in acceptance tests
+func TestBackendAcc_LoginWithCallerIdentity(t *testing.T) {
+	// This test case should be run only when certain env vars are set and
+	// executed as an acceptance test.
+	if os.Getenv(logicaltest.TestEnvVar) == "" {
+		t.Skip(fmt.Sprintf("Acceptance tests skipped unless env '%s' set", logicaltest.TestEnvVar))
+		return
+	}
+
+	storage := &logical.InmemStorage{}
+	config := logical.TestBackendConfig()
+	config.StorageView = storage
+	b, err := Backend(config)
+	if err != nil {
+		t.Fatal(err)
+	}
+	_, err = b.Setup(config)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Override the default AWS env vars (if set) with our test creds
+	// so that the credential provider chain will pick them up
+	// NOTE that I'm not bothing to override the shared config file location,
+	// so if creds are specified there, they will be used before IAM
+	// instance profile creds
+	// This doesn't provide perfect leakage protection (e.g., it will still
+	// potentially pick up credentials from the ~/.config files), but probably
+	// good enough rather than having to muck around in the low-level details
+	for _, envvar := range []string{
+		"AWS_ACCESS_KEY_ID", "AWS_SECRET_ACCESS_KEY", "AWS_SECURITY_TOKEN", "AWS_SESSION_TOKEN"} {
+		// restore existing environment variables (in case future tests need them)
+		defer os.Setenv(envvar, os.Getenv(envvar))
+		os.Setenv(envvar, os.Getenv("TEST_"+envvar))
+	}
+	awsSession, err := session.NewSession()
+	if err != nil {
+		fmt.Println("failed to create session,", err)
+		return
+	}
+
+	stsService := sts.New(awsSession)
+	stsInputParams := &sts.GetCallerIdentityInput{}
+
+	testIdentity, err := stsService.GetCallerIdentity(stsInputParams)
+	if err != nil {
+		t.Fatalf("Received error retrieving identity: %s", err)
+	}
+	testIdentityArn, _, _, err := parseIamArn(*testIdentity.Arn)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Test setup largely done
+	// At this point, we're going to:
+	// 1. Configure the client to require our test header value
+	// 2. Configure two different roles:
+	//    a. One bound to our test user
+	//    b. One bound to a garbage ARN
+	// 3. Pass in a request that doesn't have the signed header, ensure
+	//    we're not allowed to login
+	// 4. Passin a request that has a validly signed header, but the wrong
+	//    value, ensure it doesn't allow login
+	// 5. Pass in a request that has a validly signed request, ensure
+	//    it allows us to login to our role
+	// 6. Pass in a request that has a validly signed request, asking for
+	//    the other role, ensure it fails
+	const testVaultHeaderValue = "VaultAcceptanceTesting"
+	const testValidRoleName = "valid-role"
+	const testInvalidRoleName = "invalid-role"
+
+	clientConfigData := map[string]interface{}{
+		"iam_server_id_header_value": testVaultHeaderValue,
+	}
+	clientRequest := &logical.Request{
+		Operation: logical.UpdateOperation,
+		Path:      "config/client",
+		Storage:   storage,
+		Data:      clientConfigData,
+	}
+	_, err = b.HandleRequest(clientRequest)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// configuring the valid role we'll be able to login to
+	roleData := map[string]interface{}{
+		"bound_iam_principal_arn": testIdentityArn,
+		"policies":                "root",
+		"auth_type":               iamAuthType,
+	}
+	roleRequest := &logical.Request{
+		Operation: logical.CreateOperation,
+		Path:      "role/" + testValidRoleName,
+		Storage:   storage,
+		Data:      roleData,
+	}
+	resp, err := b.HandleRequest(roleRequest)
+	if err != nil || (resp != nil && resp.IsError()) {
+		t.Fatalf("bad: failed to create role: resp:%#v\nerr:%v", resp, err)
+	}
+
+	// configuring a valid role we won't be able to login to
+	roleDataEc2 := map[string]interface{}{
+		"auth_type":    "ec2",
+		"policies":     "root",
+		"bound_ami_id": "ami-1234567",
+	}
+	roleRequestEc2 := &logical.Request{
+		Operation: logical.CreateOperation,
+		Path:      "role/ec2only",
+		Storage:   storage,
+		Data:      roleDataEc2,
+	}
+	resp, err = b.HandleRequest(roleRequestEc2)
+	if err != nil || (resp != nil && resp.IsError()) {
+		t.Fatalf("bad: failed to create role; resp:%#v\nerr:%v", resp, err)
+	}
+
+	// now we're creating the invalid role we won't be able to login to
+	roleData["bound_iam_principal_arn"] = "arn:aws:iam::123456789012:role/FakeRole"
+	roleRequest.Path = "role/" + testInvalidRoleName
+	resp, err = b.HandleRequest(roleRequest)
+	if err != nil || (resp != nil && resp.IsError()) {
+		t.Fatalf("bad: didn't fail to create role: resp:%#v\nerr:%v", resp, err)
+	}
+
+	// now, create the request without the signed header
+	stsRequestNoHeader, _ := stsService.GetCallerIdentityRequest(stsInputParams)
+	stsRequestNoHeader.Sign()
+	loginData, err := buildCallerIdentityLoginData(stsRequestNoHeader.HTTPRequest, testValidRoleName)
+	if err != nil {
+		t.Fatal(err)
+	}
+	loginRequest := &logical.Request{
+		Operation: logical.UpdateOperation,
+		Path:      "login",
+		Storage:   storage,
+		Data:      loginData,
+	}
+	resp, err = b.HandleRequest(loginRequest)
+	if err != nil || resp == nil || !resp.IsError() {
+		t.Errorf("bad: expected failed login due to missing header: resp:%#v\nerr:%v", resp, err)
+	}
+
+	// create the request with the invalid header value
+
+	// Not reusing stsRequestNoHeader because the process of signing the request
+	// and reading the body modifies the underlying request, so it's just cleaner
+	// to get new requests.
+	stsRequestInvalidHeader, _ := stsService.GetCallerIdentityRequest(stsInputParams)
+	stsRequestInvalidHeader.HTTPRequest.Header.Add(iamServerIdHeader, "InvalidValue")
+	stsRequestInvalidHeader.Sign()
+	loginData, err = buildCallerIdentityLoginData(stsRequestInvalidHeader.HTTPRequest, testValidRoleName)
+	if err != nil {
+		t.Fatal(err)
+	}
+	loginRequest = &logical.Request{
+		Operation: logical.UpdateOperation,
+		Path:      "login",
+		Storage:   storage,
+		Data:      loginData,
+	}
+	resp, err = b.HandleRequest(loginRequest)
+	if err != nil || resp == nil || !resp.IsError() {
+		t.Errorf("bad: expected failed login due to invalid header: resp:%#v\nerr:%v", resp, err)
+	}
+
+	// Now, valid request against invalid role
+	stsRequestValid, _ := stsService.GetCallerIdentityRequest(stsInputParams)
+	stsRequestValid.HTTPRequest.Header.Add(iamServerIdHeader, testVaultHeaderValue)
+	stsRequestValid.Sign()
+	loginData, err = buildCallerIdentityLoginData(stsRequestValid.HTTPRequest, testInvalidRoleName)
+	if err != nil {
+		t.Fatal(err)
+	}
+	loginRequest = &logical.Request{
+		Operation: logical.UpdateOperation,
+		Path:      "login",
+		Storage:   storage,
+		Data:      loginData,
+	}
+	resp, err = b.HandleRequest(loginRequest)
+	if err != nil || resp == nil || !resp.IsError() {
+		t.Errorf("bad: expected failed login due to invalid role: resp:%#v\nerr:%v", resp, err)
+	}
+
+	loginData["role"] = "ec2only"
+	resp, err = b.HandleRequest(loginRequest)
+	if err != nil || resp == nil || !resp.IsError() {
+		t.Errorf("bad: expected failed login due to bad auth type: resp:%#v\nerr:%v", resp, err)
+	}
+
+	// finally, the happy path tests :)
+
+	loginData["role"] = testValidRoleName
+	resp, err = b.HandleRequest(loginRequest)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if resp == nil || resp.Auth == nil || resp.IsError() {
+		t.Errorf("bad: expected valid login: resp:%#v", resp)
 	}
 }