+++ builtin/credential/aws/path_role.go
@@ -1,4 +1,4 @@
-package awsec2
+package awsauth
 
 import (
 	"fmt"
@@ -20,6 +20,11 @@ func pathRole(b *backend) *framework.Path {
 				Type:        framework.TypeString,
 				Description: "Name of the role.",
 			},
+			"auth_type": {
+				Type: framework.TypeString,
+				Description: `The auth_type permitted to authenticate to this role. Must be one of
+iam or ec2 and cannot be changed after role creation.`,
+			},
 			"bound_ami_id": {
 				Type: framework.TypeString,
 				Description: `If set, defines a constraint on the EC2 instances that they should be
@@ -30,9 +35,16 @@ using the AMI ID specified by this parameter.`,
 				Description: `If set, defines a constraint on the EC2 instances that the account ID
 in its identity document to match the one specified by this parameter.`,
 			},
+			"bound_iam_principal_arn": {
+				Type: framework.TypeString,
+				Description: `ARN of the IAM principal to bind to this role. Only applicable when
+auth_type is iam.`,
+			},
 			"bound_region": {
-				Type:        framework.TypeString,
-				Description: `If set, defines a constraint on the EC2 instances that the region in its identity document to match the one specified by this parameter.`,
+				Type: framework.TypeString,
+				Description: `If set, defines a constraint on the EC2 instances that the region in
+its identity document to match the one specified by this parameter. Only applicable when
+auth_type is ec2.`,
 			},
 			"bound_iam_role_arn": {
 				Type: framework.TypeString,
@@ -41,14 +53,34 @@ that it must match the IAM role ARN specified by this parameter.
 The value is prefix-matched (as though it were a glob ending in
 '*').  The configured IAM user or EC2 instance role must be allowed
 to execute the 'iam:GetInstanceProfile' action if this is
-specified.`,
+specified. This is only checked when auth_type is
+ec2.`,
 			},
 			"bound_iam_instance_profile_arn": {
 				Type: framework.TypeString,
 				Description: `If set, defines a constraint on the EC2 instances to be associated
 with an IAM instance profile ARN which has a prefix that matches
 the value specified by this parameter. The value is prefix-matched
-(as though it were a glob ending in '*').`,
+(as though it were a glob ending in '*'). This is only checked when
+auth_type is ec2.`,
+			},
+			"inferred_entity_type": {
+				Type: framework.TypeString,
+				Description: `When auth_type is iam, the
+AWS entity type to infer from the authenticated principal. The only supported
+value is ec2_instance, which will extract the EC2 instance ID from the
+authenticated role and apply the following restrictions specific to EC2
+instances: bound_ami_id, bound_account_id, bound_iam_role_arn,
+bound_iam_instance_profile_arn, bound_vpc_id, bound_subnet_id. The configured
+EC2 client must be able to find the inferred instance ID in the results, and the
+instance must be running. If unable to determine the EC2 instance ID or unable
+to find the EC2 instance ID among running instances, then authentication will
+fail.`,
+			},
+			"inferred_aws_region": {
+				Type: framework.TypeString,
+				Description: `When auth_type is iam and
+inferred_entity_type is set, the region to assume the inferred entity exists in.`,
 			},
 			"bound_vpc_id": {
 				Type: framework.TypeString,
@@ -63,9 +95,13 @@ If set, defines a constraint on the EC2 instance to be associated with the
 subnet ID that matches the value specified by this parameter.`,
 			},
 			"role_tag": {
-				Type:        framework.TypeString,
-				Default:     "",
-				Description: "If set, enables the role tags for this role. The value set for this field should be the 'key' of the tag on the EC2 instance. The 'value' of the tag should be generated using 'role/<role>/tag' endpoint. Defaults to an empty string, meaning that role tags are disabled.",
+				Type:    framework.TypeString,
+				Default: "",
+				Description: `If set, enables the role tags for this role. The value set for this
+field should be the 'key' of the tag on the EC2 instance. The 'value'
+of the tag should be generated using 'role/<role>/tag' endpoint.
+Defaults to an empty string, meaning that role tags are disabled. This
+is only allowed if auth_type is ec2.`,
 			},
 			"period": &framework.FieldSchema{
 				Type:    framework.TypeDurationSecond,
@@ -90,9 +126,14 @@ to 0, in which case the value will fallback to the system/mount defaults.`,
 				Description: "Policies to be set on tokens issued using this role.",
 			},
 			"allow_instance_migration": {
-				Type:        framework.TypeBool,
-				Default:     false,
-				Description: "If set, allows migration of the underlying instance where the client resides. This keys off of pendingTime in the metadata document, so essentially, this disables the client nonce check whenever the instance is migrated to a new host and pendingTime is newer than the previously-remembered time. Use with caution.",
+				Type:    framework.TypeBool,
+				Default: false,
+				Description: `If set, allows migration of the underlying instance where the client
+resides. This keys off of pendingTime in the metadata document, so
+essentially, this disables the client nonce check whenever the
+instance is migrated to a new host and pendingTime is newer than the
+previously-remembered time. Use with caution. This is only checked when
+auth_type is ec2.`,
 			},
 			"disallow_reauthentication": {
 				Type:        framework.TypeBool,
@@ -159,9 +200,44 @@ func (b *backend) lockedAWSRole(s logical.Storage, roleName string) (*awsRoleEnt
 	}
 
 	b.roleMutex.RLock()
-	defer b.roleMutex.RUnlock()
-
-	return b.nonLockedAWSRole(s, roleName)
+	roleEntry, err := b.nonLockedAWSRole(s, roleName)
+	// we manually unlock rather than defer the unlock because we might need to grab
+	// a read/write lock in the upgrade path
+	b.roleMutex.RUnlock()
+	if err != nil {
+		return nil, err
+	}
+	if roleEntry == nil {
+		return nil, nil
+	}
+	needUpgrade, err := upgradeRoleEntry(roleEntry)
+	if err != nil {
+		return nil, fmt.Errorf("error upgrading roleEntry: %v", err)
+	}
+	if needUpgrade {
+		b.roleMutex.Lock()
+		defer b.roleMutex.Unlock()
+		// Now that we have a R/W lock, we need to re-read the role entry in case it was
+		// written to between releasing the read lock and acquiring the write lock
+		roleEntry, err = b.nonLockedAWSRole(s, roleName)
+		if err != nil {
+			return nil, err
+		}
+		// somebody deleted the role, so no use in putting it back
+		if roleEntry == nil {
+			return nil, nil
+		}
+		// now re-check to see if we need to upgrade
+		if needUpgrade, err = upgradeRoleEntry(roleEntry); err != nil {
+			return nil, fmt.Errorf("error upgrading roleEntry: %v", err)
+		}
+		if needUpgrade {
+			if err = b.nonLockedSetAWSRole(s, roleName, roleEntry); err != nil {
+				return nil, fmt.Errorf("error saving upgraded roleEntry: %v", err)
+			}
+		}
+	}
+	return roleEntry, nil
 }
 
 // lockedSetAWSRole creates or updates a role in the storage. This method
@@ -206,9 +282,41 @@ func (b *backend) nonLockedSetAWSRole(s logical.Storage, roleName string,
 	return nil
 }
 
+// If needed, updates the role entry and returns a bool indicating if it was updated
+// (and thus needs to be persisted)
+func upgradeRoleEntry(roleEntry *awsRoleEntry) (bool, error) {
+	if roleEntry == nil {
+		return false, fmt.Errorf("received nil roleEntry")
+	}
+	var upgraded bool
+	// Check if the value held by role ARN field is actually an instance profile ARN
+	if roleEntry.BoundIamRoleARN != "" && strings.Contains(roleEntry.BoundIamRoleARN, ":instance-profile/") {
+		// If yes, move it to the correct field
+		roleEntry.BoundIamInstanceProfileARN = roleEntry.BoundIamRoleARN
+
+		// Reset the old field
+		roleEntry.BoundIamRoleARN = ""
+
+		upgraded = true
+	}
+
+	// Check if there was no pre-existing AuthType set (from older versions)
+	if roleEntry.AuthType == "" {
+		// then default to the original behavior of ec2
+		roleEntry.AuthType = ec2AuthType
+		upgraded = true
+	}
+
+	return upgraded, nil
+
+}
+
 // nonLockedAWSRole returns the properties set on the given role. This method
 // does not acquire the read lock before reading the role from the storage. If
 // locking is desired, use lockedAWSRole instead.
+// This method also does NOT check to see if a role upgrade is required. It is
+// the responsibility of the caller to check if a role upgrade is required and,
+// if so, to upgrade the role
 func (b *backend) nonLockedAWSRole(s logical.Storage, roleName string) (*awsRoleEntry, error) {
 	if roleName == "" {
 		return nil, fmt.Errorf("missing role name")
@@ -225,20 +333,6 @@ func (b *backend) nonLockedAWSRole(s logical.Storage, roleName string) (*awsRole
 	var result awsRoleEntry
 	if err := entry.DecodeJSON(&result); err != nil {
 		return nil, err
-	}
-
-	// Check if the value held by role ARN field is actually an instance profile ARN
-	if result.BoundIamRoleARN != "" && strings.Contains(result.BoundIamRoleARN, ":instance-profile/") {
-		// If yes, move it to the correct field
-		result.BoundIamInstanceProfileARN = result.BoundIamRoleARN
-
-		// Reset the old field
-		result.BoundIamRoleARN = ""
-
-		// Save the update
-		if err = b.nonLockedSetAWSRole(s, roleName, &result); err != nil {
-			return nil, fmt.Errorf("failed to move instance profile ARN to bound_iam_instance_profile_arn field")
-		}
 	}
 
 	return &result, nil
@@ -316,6 +410,17 @@ func (b *backend) pathRoleCreateUpdate(
 	}
 	if roleEntry == nil {
 		roleEntry = &awsRoleEntry{}
+	} else {
+		needUpdate, err := upgradeRoleEntry(roleEntry)
+		if err != nil {
+			return logical.ErrorResponse(fmt.Sprintf("failed to update roleEntry: %v", err)), nil
+		}
+		if needUpdate {
+			err = b.nonLockedSetAWSRole(req.Storage, roleName, roleEntry)
+			if err != nil {
+				return logical.ErrorResponse(fmt.Sprintf("failed to save upgraded roleEntry: %v", err)), nil
+			}
+		}
 	}
 
 	// Fetch and set the bound parameters. There can't be default values
@@ -348,16 +453,120 @@ func (b *backend) pathRoleCreateUpdate(
 		roleEntry.BoundIamInstanceProfileARN = boundIamInstanceProfileARNRaw.(string)
 	}
 
-	// Ensure that at least one bound is set on the role
-	switch {
-	case roleEntry.BoundAmiID != "":
-	case roleEntry.BoundAccountID != "":
-	case roleEntry.BoundRegion != "":
-	case roleEntry.BoundVpcID != "":
-	case roleEntry.BoundSubnetID != "":
-	case roleEntry.BoundIamRoleARN != "":
-	case roleEntry.BoundIamInstanceProfileARN != "":
-	default:
+	if boundIamPrincipalARNRaw, ok := data.GetOk("bound_iam_principal_arn"); ok {
+		roleEntry.BoundIamPrincipalARN = boundIamPrincipalARNRaw.(string)
+	}
+
+	if inferRoleTypeRaw, ok := data.GetOk("inferred_entity_type"); ok {
+		roleEntry.InferredEntityType = inferRoleTypeRaw.(string)
+	}
+
+	if inferredAWSRegionRaw, ok := data.GetOk("inferred_aws_region"); ok {
+		roleEntry.InferredAWSRegion = inferredAWSRegionRaw.(string)
+	}
+
+	// auth_type is a special case as it's immutable and can't be changed once a role is created
+	if authTypeRaw, ok := data.GetOk("auth_type"); ok {
+		// roleEntry.AuthType should only be "" when it's a new role; existing roles without an
+		// auth_type should have already been upgraded to have one before we get here
+		if roleEntry.AuthType == "" {
+			switch authTypeRaw.(string) {
+			case ec2AuthType, iamAuthType:
+				roleEntry.AuthType = authTypeRaw.(string)
+			default:
+				return logical.ErrorResponse(fmt.Sprintf("unrecognized auth_type: %v", authTypeRaw.(string))), nil
+			}
+		} else if authTypeRaw.(string) != roleEntry.AuthType {
+			return logical.ErrorResponse("changing auth_type on a role is not allowed"), nil
+		}
+	} else if req.Operation == logical.CreateOperation {
+		switch req.MountType {
+		// maintain backwards compatibility for old aws-ec2 auth types
+		case "aws-ec2":
+			roleEntry.AuthType = ec2AuthType
+		// but default to iamAuth for new mounts going forward
+		case "aws":
+			roleEntry.AuthType = iamAuthType
+		default:
+			roleEntry.AuthType = iamAuthType
+		}
+	}
+
+	allowEc2Binds := roleEntry.AuthType == ec2AuthType
+
+	if roleEntry.InferredEntityType != "" {
+		switch {
+		case roleEntry.AuthType != iamAuthType:
+			return logical.ErrorResponse("specified inferred_entity_type but didn't allow iam auth_type"), nil
+		case roleEntry.InferredEntityType != ec2EntityType:
+			return logical.ErrorResponse(fmt.Sprintf("specified invalid inferred_entity_type: %s", roleEntry.InferredEntityType)), nil
+		case roleEntry.InferredAWSRegion == "":
+			return logical.ErrorResponse("specified inferred_entity_type but not inferred_aws_region"), nil
+		}
+		allowEc2Binds = true
+	} else if roleEntry.InferredAWSRegion != "" {
+		return logical.ErrorResponse("specified inferred_aws_region but not inferred_entity_type"), nil
+	}
+
+	numBinds := 0
+
+	if roleEntry.BoundAccountID != "" {
+		if !allowEc2Binds {
+			return logical.ErrorResponse(fmt.Sprintf("specified bound_account_id but not allowing ec2 auth_type or inferring %s", ec2EntityType)), nil
+		}
+		numBinds++
+	}
+
+	if roleEntry.BoundRegion != "" {
+		if roleEntry.AuthType != ec2AuthType {
+			return logical.ErrorResponse("specified bound_region but not allowing ec2 auth_type"), nil
+		}
+		numBinds++
+	}
+
+	if roleEntry.BoundAmiID != "" {
+		if !allowEc2Binds {
+			return logical.ErrorResponse(fmt.Sprintf("specified bound_ami_id but not allowing ec2 auth_type or inferring %s", ec2EntityType)), nil
+		}
+		numBinds++
+	}
+
+	if roleEntry.BoundIamInstanceProfileARN != "" {
+		if !allowEc2Binds {
+			return logical.ErrorResponse(fmt.Sprintf("specified bound_iam_instance_profile_arn but not allowing ec2 auth_type or inferring %s", ec2EntityType)), nil
+		}
+		numBinds++
+	}
+
+	if roleEntry.BoundIamRoleARN != "" {
+		if !allowEc2Binds {
+			return logical.ErrorResponse(fmt.Sprintf("specified bound_iam_role_arn but not allowing ec2 auth_type or inferring %s", ec2EntityType)), nil
+		}
+		numBinds++
+	}
+
+	if roleEntry.BoundIamPrincipalARN != "" {
+		if roleEntry.AuthType != iamAuthType {
+			return logical.ErrorResponse("specified bound_iam_principal_arn but not allowing iam auth_type"), nil
+		}
+		numBinds++
+	}
+
+	if roleEntry.BoundVpcID != "" {
+		if !allowEc2Binds {
+			return logical.ErrorResponse(fmt.Sprintf("specified bound_vpc_id but not allowing ec2 auth_type or inferring %s", ec2EntityType)), nil
+		}
+		numBinds++
+	}
+
+	if roleEntry.BoundSubnetID != "" {
+		if !allowEc2Binds {
+			return logical.ErrorResponse(fmt.Sprintf("specified bound_subnet_id but not allowing ec2 auth_type or inferring %s", ec2EntityType)), nil
+		}
+		numBinds++
+	}
+
+	if numBinds == 0 {
 		return logical.ErrorResponse("at least be one bound parameter should be specified on the role"), nil
 	}
 
@@ -370,15 +579,21 @@ func (b *backend) pathRoleCreateUpdate(
 
 	disallowReauthenticationBool, ok := data.GetOk("disallow_reauthentication")
 	if ok {
+		if roleEntry.AuthType != ec2AuthType {
+			return logical.ErrorResponse("specified disallow_reauthentication when not using ec2 auth type"), nil
+		}
 		roleEntry.DisallowReauthentication = disallowReauthenticationBool.(bool)
-	} else if req.Operation == logical.CreateOperation {
+	} else if req.Operation == logical.CreateOperation && roleEntry.AuthType == ec2AuthType {
 		roleEntry.DisallowReauthentication = data.Get("disallow_reauthentication").(bool)
 	}
 
 	allowInstanceMigrationBool, ok := data.GetOk("allow_instance_migration")
 	if ok {
+		if roleEntry.AuthType != ec2AuthType {
+			return logical.ErrorResponse("specified allow_instance_migration when not using ec2 auth type"), nil
+		}
 		roleEntry.AllowInstanceMigration = allowInstanceMigrationBool.(bool)
-	} else if req.Operation == logical.CreateOperation {
+	} else if req.Operation == logical.CreateOperation && roleEntry.AuthType == ec2AuthType {
 		roleEntry.AllowInstanceMigration = data.Get("allow_instance_migration").(bool)
 	}
 
@@ -430,13 +645,16 @@ func (b *backend) pathRoleCreateUpdate(
 
 	roleTagStr, ok := data.GetOk("role_tag")
 	if ok {
+		if roleEntry.AuthType != ec2AuthType {
+			return logical.ErrorResponse("tried to enable role_tag when not using ec2 auth method"), nil
+		}
 		roleEntry.RoleTag = roleTagStr.(string)
 		// There is a limit of 127 characters on the tag key for AWS EC2 instances.
 		// Complying to that requirement, do not allow the value of 'key' to be more than that.
 		if len(roleEntry.RoleTag) > 127 {
 			return logical.ErrorResponse("length of role tag exceeds the EC2 key limit of 127 characters"), nil
 		}
-	} else if req.Operation == logical.CreateOperation {
+	} else if req.Operation == logical.CreateOperation && roleEntry.AuthType == ec2AuthType {
 		roleEntry.RoleTag = data.Get("role_tag").(string)
 	}
 
@@ -460,13 +678,17 @@ func (b *backend) pathRoleCreateUpdate(
 
 // Struct to hold the information associated with an AMI ID in Vault.
 type awsRoleEntry struct {
+	AuthType                   string        `json:"auth_type" structs:"auth_type" mapstructure:"auth_type"`
 	BoundAmiID                 string        `json:"bound_ami_id" structs:"bound_ami_id" mapstructure:"bound_ami_id"`
 	BoundAccountID             string        `json:"bound_account_id" structs:"bound_account_id" mapstructure:"bound_account_id"`
-	BoundRegion                string        `json:"bound_region" structs:"bound_region" mapstructure:"bound_region"`
+	BoundIamPrincipalARN       string        `json:"bound_iam_principal_arn" structs:"bound_iam_principal_arn" mapstructure:"bound_iam_principal_arn"`
 	BoundIamRoleARN            string        `json:"bound_iam_role_arn" structs:"bound_iam_role_arn" mapstructure:"bound_iam_role_arn"`
 	BoundIamInstanceProfileARN string        `json:"bound_iam_instance_profile_arn" structs:"bound_iam_instance_profile_arn" mapstructure:"bound_iam_instance_profile_arn"`
+	BoundRegion                string        `json:"bound_region" structs:"bound_region" mapstructure:"bound_region"`
 	BoundSubnetID              string        `json:"bound_subnet_id" structs:"bound_subnet_id" mapstructure:"bound_subnet_id"`
 	BoundVpcID                 string        `json:"bound_vpc_id" structs:"bound_vpc_id" mapstructure:"bound_vpc_id"`
+	InferredEntityType         string        `json:"inferred_entity_type" structs:"inferred_entity_type" mapstructure:"inferred_entity_type"`
+	InferredAWSRegion          string        `json:"inferred_aws_region" structs:"inferred_aws_region" mapstructure:"inferred_aws_region"`
 	RoleTag                    string        `json:"role_tag" structs:"role_tag" mapstructure:"role_tag"`
 	AllowInstanceMigration     bool          `json:"allow_instance_migration" structs:"allow_instance_migration" mapstructure:"allow_instance_migration"`
 	TTL                        time.Duration `json:"ttl" structs:"ttl" mapstructure:"ttl"`
@@ -494,6 +716,7 @@ endpoint 'role/<role>/tag'. This tag then needs to be applied on the
 instance before it attempts a login. The policies on the tag should be a
 subset of policies that are associated to the role. In order to enable
 login using tags, 'role_tag' option should be set while creating a role.
+This only applies when authenticating EC2 instances.
 
 Also, a 'max_ttl' can be configured in this endpoint that determines the maximum
 duration for which a login can be renewed. Note that the 'max_ttl' has an upper