+++ vault/router.go
@@ -257,6 +257,7 @@ func (r *Router) routeCommon(req *logical.Request, existenceCheck bool) (*logica
 	originalPath := req.Path
 	req.Path = strings.TrimPrefix(req.Path, mount)
 	req.MountPoint = mount
+	req.MountType = re.mountEntry.Type
 	if req.Path == "/" {
 		req.Path = ""
 	}
@@ -303,7 +304,8 @@ func (r *Router) routeCommon(req *logical.Request, existenceCheck bool) (*logica
 	// Reset the request before returning
 	defer func() {
 		req.Path = originalPath
-		req.MountPoint = ""
+		req.MountPoint = mount
+		req.MountType = re.mountEntry.Type
 		req.Connection = originalConn
 		req.ID = originalReqID
 		req.Storage = nil